# service/src/main/resources/api/service_openapi.yaml
openapi: 3.0.3
info:
  title: Tanagra Service API
  description: github.com/DataBiosphere/tanagra
  version: 0.0.0
servers:
  - url: /
security:
  - bearerAuth: []
  - authorization: [openid, email, profile]
paths:
...
  "/v2/studies/{studyId}/cohorts":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
    get:
      parameters:
        - $ref: "#/components/parameters/OffsetV2"
        - $ref: "#/components/parameters/LimitV2"
      summary: List all cohorts in a study
      operationId: listCohorts
      tags: [CohortsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CohortListV2"
        500:
          $ref: "#/components/responses/ServerError"
    post:
      summary: Create a new cohort
      operationId: createCohort
      tags: [CohortsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CohortCreateInfoV2"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CohortV2"
        500:
          $ref: "#/components/responses/ServerError"
  "/v2/studies/{studyId}/cohorts/{cohortId}":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
    get:
      summary: Get an existing cohort
      operationId: getCohort
      tags: [CohortsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohortV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update an existing cohort
      operationId: updateCohort
      tags: [CohortsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohortUpdateInfoV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohortV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete a cohort
      operationId: deleteCohort
      tags: [CohortsV2]
      responses:
        204:
          description: OK
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
...
  "/v2/studies/{studyId}/cohorts/{cohortId}/annotations":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
    get:
      parameters:
        - $ref: "#/components/parameters/OffsetV2"
        - $ref: "#/components/parameters/LimitV2"
      summary: List all annotations for a cohort
      operationId: listAnnotations
      tags: [AnnotationsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnotationListV2"
        500:
          $ref: "#/components/responses/ServerError"
    post:
      summary: Create a new annotation
      operationId: createAnnotation
      tags: [AnnotationsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationCreateInfoV2"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnotationV2"
        500:
          $ref: "#/components/responses/ServerError"
  "/v2/studies/{studyId}/cohorts/{cohortId}/annotationValues/export":
    parameters:
    - $ref: '#/components/parameters/StudyIdV2'
    - $ref: '#/components/parameters/CohortIdV2'
    post:
      summary: Export annotation values
      description: |
        Export a table with column headers: primary entity instance id, all
        annotation keys. Table cells contain annotation values. For each cell,
        the annotation value from the latest review is used, ordered by review
        creation date. Exports to a TSV file in bucket specified in
        application.yaml.
      operationId: exportAnnotationValues
      tags: [AnnotationsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportFile"
        500:
          $ref: "#/components/responses/ServerError"
  "/v2/studies/{studyId}/cohorts/{cohortId}/annotations/{annotationId}":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
      - $ref: '#/components/parameters/AnnotationIdV2'
    get:
      summary: Get an existing annotation
      operationId: getAnnotation
      tags: [AnnotationsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotationV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update an existing annotation
      operationId: updateAnnotation
      tags: [AnnotationsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnotationUpdateInfoV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotationV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete an annotation
      operationId: deleteAnnotation
      tags: [AnnotationsV2]
      responses:
        204:
          description: OK
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  "/v2/studies/{studyId}/cohorts/{cohortId}/annotations/{annotationId}/reviews/{reviewId}/values/{valueId}":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
      - $ref: '#/components/parameters/AnnotationIdV2'
      - $ref: '#/components/parameters/ReviewIdV2'
      - $ref: '#/components/parameters/AnnotationValueIdV2'
    post:
      summary: Create or update an annotation value
      operationId: createUpdateAnnotationValue
      tags: [AnnotationsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LiteralV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotationValueV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete an annotation value
      operationId: deleteAnnotationValue
      tags: [AnnotationsV2]
      responses:
        204:
          description: OK
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
...
    StudyIdV2:
      name: studyId
      in: path
      description: ID of the study
      required: true
      schema:
        type: string
    CohortIdV2:
      name: cohortId
      in: path
      description: ID of the cohort
      required: true
      schema:
        type: string
...
    ReviewIdV2:
      name: reviewId
      in: path
      description: ID of the review
      required: true
      schema:
        type: string
    AnnotationIdV2:
      name: annotationId
      in: path
      description: ID of the annotation
      required: true
      schema:
        type: string
    AnnotationValueIdV2:
      name: valueId
      in: path
      description: ID of the annotation value
      required: true
      schema:
        type: string
    OffsetV2:
      name: offset
      in: query
      description: The number of items to skip before starting to collect the result set.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    LimitV2:
      name: limit
      in: query
      description: The maximum number of items to return. Default 15
      required: false
      schema:
        type: integer
        minimum: 1
        default: 15
  responses:
    NotFound:
      description: Not found (or unauthorized)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReportV2'
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReportV2'
  schemas:
...
    LiteralV2:
      type: object
      description: Union of references to each value type. Exactly one will be populated based on the dataType.
      nullable: true
      properties:
        dataType:
          $ref: "#/components/schemas/DataTypeV2"
        valueUnion:
          type: object
          properties:
            boolVal:
              type: boolean
            int64Val:
              type: integer
              format: int64
            stringVal:
              type: string
            dateVal:
              type: string
              description: Format YYYY-MM-DD
    DataTypeV2:
      type: string
      enum: ["INT64", "STRING", "BOOLEAN", "DATE"]
...
    CohortV2:
      type: object
      properties:
        id:
          description: ID of the cohort, immutable
          type: string
        underlayName:
          $ref: "#/components/schemas/UnderlayNameV2"
        displayName:
          $ref: "#/components/schemas/CohortDisplayNameV2"
        description:
          $ref: "#/components/schemas/CohortDescriptionV2"
        criteriaGroupSections:
          type: array
          description: Criteria group sections that define the entity filter
          items:
            $ref: "#/components/schemas/CriteriaGroupSectionV3"
        created:
          description: Timestamp of when the cohort was created
          type: string
          format: date-time
        createdBy:
          description: Email of user who created cohort
          type: string
        lastModified:
          description: Timestamp of when the cohort was last modified
          type: string
          format: date-time
      required:
        - id
        - underlayName
        - displayName
        - criteriaGroups
        - created
        - createdBy
        - lastModified
    CohortListV2:
      type: array
      items:
        $ref: "#/components/schemas/CohortV2"
    CohortCreateInfoV2:
      type: object
      properties:
        underlayName:
          $ref: "#/components/schemas/UnderlayNameV2"
        displayName:
          $ref: "#/components/schemas/CohortDisplayNameV2"
        description:
          $ref: "#/components/schemas/CohortDescriptionV2"
    CohortUpdateInfoV2:
      type: object
      properties:
        displayName:
          $ref: "#/components/schemas/CohortDisplayNameV2"
        description:
          $ref: "#/components/schemas/CohortDescriptionV2"
        criteriaGroupSections:
          type: array
          description: Criteria group sections that define the entity filter
          items:
            $ref: "#/components/schemas/CriteriaGroupSectionV3"
    CohortDisplayNameV2:
      type: string
      description: Human readable name of the cohort
    CohortDescriptionV2:
      type: string
      description: Description of the cohort
...
    UnderlayNameV2:
      type: string
      description: Name of the underlay, immutable
...
    CriteriaGroupSectionV3:
      type: object
      description: Section of criteria groups and boolean logic operators to combine them
      properties:
        id:
          type: string
          description: ID of the section, immutable
        displayName:
          type: string
          description: Name of the section
        criteriaGroups:
          type: array
          description: Set of criteria groups in the section
          items:
            $ref: "#/components/schemas/CriteriaGroupV3"
        operator:
          type: string
          description: Operator to use when combining the criteria groups in the section
          enum: ["AND", "OR"]
        excluded:
          type: boolean
          description: True to exclude the section, false to include it
      required:
        - id
        - displayName
        - criteriaGroups
        - operator
        - excluded
    CriteriaGroupV3:
      type: object
      description: Group of related criteria (primary + modifiers)
      properties:
        id:
          type: string
          description: ID of the group, immutable
        displayName:
          type: string
          description: Name of the group
        criteria:
          type: array
          description: Set of criteria in the group
          items:
            $ref: "#/components/schemas/CriteriaV2"
        entity:
          description: Entity referenced by the criteria
          type: string
        groupByCountOperator:
          description: Operator to use for the optional group by count modifier
          $ref: "#/components/schemas/BinaryOperatorV2"
        groupByCountValue:
          type: integer
          description: Value to use for the optional group by count modifier
          nullable: true
      required:
        - id
        - displayName
        - criteria
        - entity
    CriteriaV2:
      type: object
      description: Single criteria for a cohort or concept set definition
      properties:
        id:
          type: string
          description: ID of the criteria, immutable
        displayName:
          type: string
          description: Name of the criteria
        pluginName:
          type: string
          description: Name of the plugin that generated this criteria
        selectionData:
          type: string # JSON formatted
          description: Serialized plugin-specific representation of the user's selection
        uiConfig:
          # TODO: Remove this uiConfig property from the service API once we fully support UI plugins, because it
          # should be identical to the static config stored in the underlay. For now, we save it separately for
          # each criteria so we can handle schema changes in the UI code instead of in plugin-specific backend code.
          type: string # JSON formatted
          description: Serialized plugin-specific UI configuration for the criteria
        tags:
          type: array
          description: Unstructured tags for the criteria
          items:
            type: string
      required:
        - id
        - displayName
        - pluginName
        - selectionData
        - uiConfig
...
    ReviewIdV2:
      type: string
      description: ID of the review, immutable
...
    AnnotationV2:
      type: object
      properties:
        id:
          description: ID of the annotation, immutable
          type: string
        displayName:
          $ref: "#/components/schemas/AnnotationDisplayNameV2"
        description:
          $ref: "#/components/schemas/AnnotationDescriptionV2"
        dataType:
          $ref: "#/components/schemas/DataTypeV2"
        enumVals:
          type: array
          description: |
            May only be set if dataType=STRING. Annotation value must be one of
            enumVals.
          items:
            type: string
      required:
        - id
        - displayName
        - dataType
    AnnotationListV2:
      type: array
      items:
        $ref: "#/components/schemas/AnnotationV2"
    AnnotationCreateInfoV2:
      type: object
      properties:
        displayName:
          $ref: "#/components/schemas/AnnotationDisplayNameV2"
        description:
          $ref: "#/components/schemas/AnnotationDescriptionV2"
        dataType:
          $ref: "#/components/schemas/DataTypeV2"
        enumVals:
          type: array
          description: |
            May only be set if dataType=STRING. Annotation value must be one of
            enumVals.
          items:
            type: string
      required:
        - displayName
        - dataType
    AnnotationUpdateInfoV2:
      type: object
      properties:
        displayName:
          $ref: "#/components/schemas/AnnotationDisplayNameV2"
        description:
          $ref: "#/components/schemas/AnnotationDescriptionV2"
    AnnotationDisplayNameV2:
      type: string
      description: Human readable name of the annotation
    AnnotationDescriptionV2:
      type: string
      description: Description of the annotation
    AnnotationValueV2:
      type: object
      properties:
        id:
          description: ID of the annotation value, immutable
          type: string
        value:
          $ref: "#/components/schemas/LiteralV2"
        review:
          $ref: "#/components/schemas/ReviewIdV2"
      required:
        - id
        - value
        - review
...
    ErrorReportV2:
      type: object
      required: [message, statusCode, causes]
      properties:
        message:
          type: string
        statusCode:
          type: integer
        causes:
          type: array
          items:
            type: string
    ExportFile:
      type: object
      properties:
        gcsSignedUrl:
          description: GCS signed URL of exported dataset CSV file
          type: string
    BinaryOperatorV2:
      type: string
      enum: ["EQUALS", "NOT_EQUALS", "LESS_THAN", "GREATER_THAN"]
  securitySchemes:
    authorization:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization
    bearerAuth:
      type: http
      scheme: bearer